requirements.txt
python-dotenv
chainlit
langchain
langchain-openai
langchain-community